Clarify PTC content
My understanding is that PTC practice items are meant to be statements of qualitative improvement that have either been made already or are being undertaken, not actions to be undertaken. So an action might be "Team develops triggers and scripts to run the tests" but the qualitative improvement would be "Tests are run automatically according to the teams policy". The former is an action that needs to be carried out, but the latter is a statement that demonstrates improvement from some previous situation.